LayoutTests/imported/w3c/web-platform-tests/hr-time/performance-tojson-expected.txt


FAIL Test performance.toJSON() assert_equals: expected "function" but got "undefined"
